fix: update all return types in CGO functions from C.int to C.uint64_t
- Updated return types for various CGO functions in unary, streaming, and common headers to use C.uint64_t instead of C.int for better error handling.
- Modified the corresponding function signatures in the generated code to reflect the new return type.
- Adjusted error handling to return the error ID as a uint64_t, ensuring consistency across the codebase.

+++ cgotest/c_tests/ygrpc_cgo_common.h
@@ -16,8 +16,8 @@ typedef enum {
 } YgrpcProtocol;
 
 extern void Ygrpc_Free(void* ptr);
-extern int Ygrpc_SetProtocol(int protocol);
-extern int Ygrpc_GetErrorMsg(uint64_t error_id, void** msg_ptr, int* msg_len, FreeFunc* msg_free);
+extern uint64_t Ygrpc_SetProtocol(int protocol);
+extern uint64_t Ygrpc_GetErrorMsg(uint64_t error_id, void** msg_ptr, int* msg_len, FreeFunc* msg_free);
 
 static inline void call_free_func(FreeFunc fn, void* ptr) {
     if (fn) fn(ptr);
